Rename ._crs to just .crs and exclude some tools from flat namespace
